cia_3/requirements.txt
fastapi
uvicorn
python-multipart
inference-sdk
pillow
jinja2
httpx
python-dotenv


# uvicorn app:app --reload